If one model is has_manied to multiple other models and they had the same association name (and why wouldn't they) the first association to be called would be the only one you could use without causing a type mismatch error because of the caching we did not being specific enough.
E.g.

Bone.belongs_to :dog
Bone.belongs_to :skeleton
Dog.has_many :bones
Skeleton.has_many :bones

If we started with Bone and asked for dog then asked for skeleton, we'd get a mismatch on trying to set a Skeleton into the Dog.  We now make sure the reflection name and the instance class are used to cache the found association.