[#4] Trap NameError's that occur when using reflection.klass on polymorphics
The plugin still doesn't add any reciprocals for polymorphic associations, but it won't throw a NameError when trying to load a polymorphic association.  Generally speaking the .klass of a polymorphic association isn't actually a ruby class, just a convenient name to refer to all the types of object that can be associated with that model.