[#1] Support for setting reciprocals on direct assignment of has_one/belongs_to.
Use replace as it gets called for build and create (and create! on has_one) as well as being what <association_name>= will use.  For has_one we still need to override new_record as it can be called with replace_existing = false, and thus replace won't be called.  Have optimised for that though.